requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
h11==0.14.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
pycparser==2.21
PySocks==1.7.1
pytest==7.4.2
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-selenium==4.1.0
pytest-variables==3.1.0
requests==2.31.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
tenacity==8.2.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.0.6
wsproto==1.2.0
